cleaned up existing implementation
