+++ Assets/PickUps/Models/Hambuger.FBX.meta
