+++ Assets/PickUps/Models/Materials/mug-metal.mat.meta
