+++ Assets/PickUps/Models/Materials/mug.mat.meta
