+++ Assets/PickUps/Models/Mug.fbx
