+++ Assets/PickUps/Models/Mug.fbx.meta
