+++ Assets/PickUps/Models/Textures.meta
