+++ Assets/PickUps/Models/Textures/Hambuger.png.meta
