+++ Assets/PickUps/Models/Textures/mug-albedo-dirt.tga.meta
