+++ Assets/PickUps/Models/Textures/mug-ao.tga.meta
