+++ Assets/PickUps/Models/Textures/mug-metal.tga.meta
