+++ Assets/PickUps/Models/Textures/mug-normal.tga.meta
