+++ Assets/PickUps/Models/Textures/mug_albedo.tga.meta
