+++ Assets/PickUps/Prefabs.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d7d9b99f49f1ac34d9396d47f962ffaf
+folderAsset: yes
+timeCreated: 1517500427
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
