+++ Assets/PickUps/Prefabs/Coffee.prefab
@@ -19,17 +19,45 @@ GameObject:
   serializedVersion: 5
   m_Component:
   - component: {fileID: 4274917945014014}
-  - component: {fileID: 33635547200615036}
-  - component: {fileID: 23151460531722700}
-  - component: {fileID: 114132355675101800}
   - component: {fileID: 136401091654846760}
+  - component: {fileID: 114132355675101800}
   m_Layer: 0
-  m_Name: Pick Up Coffee
+  m_Name: Coffee
   m_TagString: Pick Up
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
+--- !u!1 &1757660261986016
+GameObject:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  serializedVersion: 5
+  m_Component:
+  - component: {fileID: 4253062129300030}
+  - component: {fileID: 33195454779062090}
+  - component: {fileID: 23537639778462068}
+  m_Layer: 0
+  m_Name: Mesh
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &4253062129300030
+Transform:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1757660261986016}
+  m_LocalRotation: {x: 0.08715578, y: 0, z: 0, w: 0.9961947}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 4274917945014014}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 10, y: 0, z: 0}
 --- !u!4 &4274917945014014
 Transform:
   m_ObjectHideFlags: 1
@@ -37,18 +65,19 @@ Transform:
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1172670757554762}
   m_LocalRotation: {x: 0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -156.81462, y: 4.5, z: -233.42355}
+  m_LocalPosition: {x: -0.09055848, y: 0.19366533, z: 3.7358747}
   m_LocalScale: {x: 4, y: 4, z: 4}
-  m_Children: []
+  m_Children:
+  - {fileID: 4253062129300030}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!23 &23151460531722700
+--- !u!23 &23537639778462068
 MeshRenderer:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1172670757554762}
+  m_GameObject: {fileID: 1757660261986016}
   m_Enabled: 1
   m_CastShadows: 1
   m_ReceiveShadows: 1
@@ -73,13 +102,14 @@ MeshRenderer:
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: 0
+  m_SortingLayer: 0
   m_SortingOrder: 0
---- !u!33 &33635547200615036
+--- !u!33 &33195454779062090
 MeshFilter:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1172670757554762}
+  m_GameObject: {fileID: 1757660261986016}
   m_Mesh: {fileID: 4300000, guid: 4e6db54670ea84d4298ec673c60ca3c7, type: 3}
 --- !u!114 &114132355675101800
 MonoBehaviour:
@@ -92,7 +122,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 86e8139601b524f4f9cc58f8c0e2cc03, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  PickupType: 
+  rotationSpeed: 60
+  oscillationMax: 0.04
+  oscillationSpeed: 1
+  currentOscillation: 0
 --- !u!136 &136401091654846760
 CapsuleCollider:
   m_ObjectHideFlags: 1
