+++ Assets/PickUps/Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: cae233db65a47e041869f6d24c05dfa1
+folderAsset: yes
+timeCreated: 1517500481
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
